refactor: 결과를 판단하는 로직을 Dealer와 Player가 스스로 담당하도록 리팩토링
