+++ src/main/java/blakcjack/domain/participant/Dealer.java
@@ -2,9 +2,12 @@ package blakcjack.domain.participant;
 
 import blakcjack.domain.card.Card;
 import blakcjack.domain.name.Name;
+import blakcjack.domain.outcome.Outcome;
 
 import java.util.Collections;
 import java.util.List;
+
+import static blakcjack.domain.outcome.Outcome.*;
 
 public class Dealer extends Participant {
 	public static final String DEALER_NAME = "딜러";
@@ -21,5 +24,25 @@ public class Dealer extends Participant {
 
 	public boolean isScoreLowerThanMaximumDrawCriterion() {
 		return calculateScore() < DEALER_MAXIMUM_DRAW_CRITERION;
+	}
+
+	public Outcome judgeOutcomeOf(final Player player) {
+		if (hasAnyBust(this, player)) {
+			return player.judgeOutcomeByBust();
+		}
+		return judgeOutcomeByScore(player);
+	}
+
+	private Outcome judgeOutcomeByScore(final Player player) {
+		final int playerScore = player.calculateScore();
+		final int dealerScore = this.calculateScore();
+
+		if (playerScore > dealerScore) {
+			return WIN;
+		}
+		if (playerScore < dealerScore) {
+			return LOSE;
+		}
+		return DRAW;
 	}
 }
